Re-organized trait tests
