+++ deep_causality/tests/traits/generatable/mod.rs
@@ -3,4 +3,4 @@
  * Copyright (c) "2025" . The DeepCausality Authors and Contributors. All Rights Reserved.
  */
 
-mod generative_processor_tests;
+mod generative_processor;
